@@ -1,36 +1,47 @@
 [tool.poetry]
+package-mode = false
 name = "qq-github-bot"
-version = "0.2.0"
+version = "3.0.4"
 description = "GitHub Bot for QQ"
-authors = ["yanyongyu <yanyongyu_1@126.com>"]
+authors = ["yanyongyu <yyy@yyydl.top>"]
 license = "MIT"
+readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.10"
-psutil = "^5.7.2"
+python = "^3.11"
 Jinja2 = "^3.0.0"
-aerich = "^0.6.3"
 unidiff = "^0.7.0"
 humanize = "^4.0.0"
-sentry-sdk = "^1.0.0"
+asyncpg = "^0.29.0"
 playwright = "^1.17.2"
 mdit-py-emoji = "^0.1.0"
-nonebot2 = "^2.0.0-rc.1"
-typing-extensions = "^4.0.0"
-nonebot-adapter-github = "^0.2.0"
-nonebot-adapter-onebot = "^2.0.0-beta.1"
-redis = { version = "^4.0.0", extras = ["hiredis"] }
-tortoise-orm = { version = "^0.19.2", extras = ["asyncpg"] }
-markdown-it-py = { version = "^2.1.0", extras = ["linkify", "plugins"] }
+nonebot-plugin-orm = "^0.7.0"
+nonebot-adapter-qq = "^1.4.2"
+nonebot-plugin-sentry = "^2.0.0"
+nonebot-plugin-status = "^0.9.0"
+nonebot-adapter-github = "^0.4.0"
+nonebot-adapter-onebot = "^2.4.2"
+# pydantic 2.9.0 handle the namedtuple type incorrectly
+# https://github.com/pydantic/pydantic/issues/10336
+pydantic = ">=2.6.3, <3.0.0, !=2.9.0"
+redis = { version = "^5.0.0", extras = ["hiredis"] }
+markdown-it-py = { version = "^3.0.0", extras = ["linkify", "plugins"] }
+nonebot2 = { version = "^2.2.0", extras = ["httpx", "websockets", "fastapi"] }
 
 [tool.poetry.group.dev.dependencies]
+ruff = "^0.6.0"
 isort = "^5.9.3"
-black = "^22.1.0"
-nonebug = "^0.2.0"
+black = "^24.0.0"
 nonemoji = "^0.1.2"
-pre-commit = "^2.20.0"
+pre-commit = "^3.1.0"
+
+[tool.poetry.group.deploy.dependencies]
+gunicorn = "^23.0.0"
+uvicorn = ">=0.18.3,<1.0.0"
 
 [tool.black]
+preview = true
+target-version = ["py311", "py312"]
 line-length = 88
 extend-exclude = '''
 '''
@@ -42,6 +53,48 @@ length_sort = true
 skip_gitignore = true
 force_sort_within_sections = true
 extra_standard_library = "typing_extensions"
+
+[tool.ruff]
+line-length = 88
+target-version = "py311"
+
+[tool.ruff.lint]
+select = [
+  "F",     # Pyflakes
+  "W",     # pycodestyle warnings
+  "E",     # pycodestyle errors
+  "UP",    # pyupgrade
+  "ASYNC", # flake8-async
+  "C4",    # flake8-comprehensions
+  "T10",   # flake8-debugger
+  "T20",   # flake8-print
+  "PYI",   # flake8-pyi
+  "PT",    # flake8-pytest-style
+  "Q",     # flake8-quotes
+  "RUF",   # Ruff-specific rules
+]
+ignore = [
+  "E402",   # module-import-not-at-top-of-file
+  "UP037",  # quoted-annotation
+  "RUF001", # ambiguous-unicode-character-string
+  "RUF002", # ambiguous-unicode-character-docstring
+  "RUF003", # ambiguous-unicode-character-comment
+]
+
+[tool.ruff.lint.extend-per-file-ignores]
+"./scripts/*" = ["T201"]
+"./docker/*" = ["T201"]
+
+[tool.pyright]
+extraPaths = ["./"]
+
+pythonPlatform = "All"
+pythonVersion = "3.11"
+
+typeCheckingMode = "standard"
+reportShadowedImports = false
+disableBytesTypePromotions = true
+reportIncompatibleVariableOverride = false
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
