@@ -1,11 +1,10 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """
 @Author         : yanyongyu
 @Date           : 2021-03-09 16:14:00
 @LastEditors    : yanyongyu
 @LastEditTime   : 2021-03-09 16:32:54
-@Description    : None
+@Description    : Foo module
 @GitHub         : https://github.com/yanyongyu
 """
+
 __author__ = "yanyongyu"
