@@ -1,47 +1,46 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """
 @Author         : yanyongyu
 @Date           : 2021-03-09 16:30:16
 @LastEditors    : yanyongyu
-@LastEditTime   : 2022-10-05 07:04:36
-@Description    : None
+@LastEditTime   : 2024-05-16 18:03:01
+@Description    : OAuth lib
 @GitHub         : https://github.com/yanyongyu
 """
+
 __author__ = "yanyongyu"
 
-import json
 import urllib.parse
 
 from src.plugins.github import config
 from src.plugins.github.models import User
+from src.providers.platform import UserInfo
 from src.plugins.github.utils import get_github
 from src.plugins.github.cache import get_state, create_state, delete_state
 
-from .platform import UserInfo, create_or_update_user
-
 
 async def create_auth_link(info: UserInfo) -> str:
+    """Create oauth link"""
     query = {
         "client_id": config.github_app.client_id,
-        "state": await create_state(json.dumps(info)),
+        "state": await create_state(info),
     }
     return f"https://github.com/login/oauth/authorize?{urllib.parse.urlencode(query)}"
 
 
-async def get_state_data(state_id: str) -> UserInfo | None:
-    return json.loads(data) if (data := await get_state(state_id)) is not None else None
-
-
-async def delete_state_data(state_id: str) -> None:
-    return await delete_state(state_id)
+async def consume_state(state_id: str) -> UserInfo | None:
+    """Get oauth state data"""
+    if user := await get_state(state_id):
+        await delete_state(state_id)
+        return user
 
 
 async def create_auth_user(info: UserInfo, access_token: str) -> User:
-    return await create_or_update_user(info, access_token=access_token)
+    """Create oauth user model"""
+    return await User.create_or_update_by_info(info, access_token=access_token)
 
 
-async def get_token_by_code(code: str) -> str:
+async def get_token_by_code(code: str) -> str | None:
+    """Get oauth token by oauth code"""
     github = get_github()
     data = {
         "client_id": config.github_app.client_id,
@@ -54,5 +53,6 @@ async def get_token_by_code(code: str) -> str:
         "https://github.com/login/oauth/access_token",
         json=data,
         headers=headers,
-    )
-    return response.json()["access_token"]
+    )  # type: ignore
+    data = response.json()
+    return data.get("access_token")
