@@ -1,50 +1,46 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """
 @Author         : yanyongyu
 @Date           : 2021-05-14 17:09:12
 @LastEditors    : yanyongyu
-@LastEditTime   : 2022-10-05 06:34:43
-@Description    : None
+@LastEditTime   : 2024-05-31 11:22:49
+@Description    : GitHub html renderer
 @GitHub         : https://github.com/yanyongyu
 """
+
 __author__ = "yanyongyu"
 
 from pathlib import Path
 from typing import Literal
 
 import jinja2
-from githubkit.rest import Issue
 
-from src.plugins.github.utils import get_github
-
+from .globals import scale_linear
 from .filters import (
     debug_event,
     markdown_gfm,
     review_state,
+    left_truncate,
     relative_time,
     markdown_emoji,
     markdown_title,
 )
-from .globals import (
-    REACTION_EMOJIS,
-    scale_linear,
-    get_issue_repo,
-    get_pull_request,
-    get_issue_timeline,
-    find_dismissed_review,
-    get_comment_reactions,
-    get_issue_label_color,
-    get_pull_request_diff,
+from .context import (
+    DiffContext,
+    IssueContext,
+    ReadmeContext,
+    IssueClosedContext,
+    IssueOpenedContext,
+    IssueCommentedContext,
+    UserContributionContext,
 )
 
 env = jinja2.Environment(
     trim_blocks=True,
     lstrip_blocks=True,
-    extensions=["jinja2.ext.loopcontrols"],
     loader=jinja2.FileSystemLoader(Path(__file__).parent / "templates"),
     enable_async=True,
 )
+"""Jinja environment for rendering"""
 
 env.filters["markdown_title"] = markdown_title
 env.filters["markdown_emoji"] = markdown_emoji
@@ -52,25 +48,97 @@ env.filters["markdown_gfm"] = markdown_gfm
 env.filters["relative_time"] = relative_time
 env.filters["debug_event"] = debug_event
 env.filters["review_state"] = review_state
+env.filters["left_truncate"] = left_truncate
 
-env.globals["get_issue_repo"] = get_issue_repo
-env.globals["get_issue_timeline"] = get_issue_timeline
-env.globals["get_pull_request"] = get_pull_request
-env.globals["get_pull_request_diff"] = get_pull_request_diff
-env.globals["get_comment_reactions"] = get_comment_reactions
-env.globals["REACTION_EMOJIS"] = REACTION_EMOJIS
-env.globals["get_issue_label_color"] = get_issue_label_color
-env.globals["find_dismissed_review"] = find_dismissed_review
 env.globals["scale_linear"] = scale_linear
 
 
-async def issue_to_html(issue: Issue, theme: Literal["light", "dark"] = "light") -> str:
+async def user_contribution_to_html(
+    ctx: UserContributionContext, theme: Literal["light", "dark"] = "light"
+) -> str:
+    """Render user contribution to image
+
+    Args:
+        ctx: the user contribution context
+        theme: the theme of the image
+    """
+    template = env.get_template("views/contribution.html.jinja")
+    return await template.render_async(ctx=ctx, theme=theme)
+
+
+async def readme_to_html(
+    ctx: ReadmeContext, theme: Literal["light", "dark"] = "light"
+) -> str:
+    """Render repo readme to html
+
+    Args:
+        ctx: the readme context
+        theme: the theme of the html
+    """
+    template = env.get_template("views/readme.html.jinja")
+    return await template.render_async(ctx=ctx, theme=theme)
+
+
+async def issue_to_html(
+    ctx: IssueContext, theme: Literal["light", "dark"] = "light"
+) -> str:
+    """Render issue or pr with timeline to html
+
+    Args:
+        ctx: the issue context
+        theme: the theme of the html
+    """
     template = env.get_template("views/issue.html.jinja")
-    return await template.render_async(issue=issue, theme=theme)
+    return await template.render_async(ctx=ctx, theme=theme)
 
 
 async def pr_diff_to_html(
-    issue: Issue, theme: Literal["light", "dark"] = "light"
+    ctx: DiffContext, theme: Literal["light", "dark"] = "light"
 ) -> str:
+    """Render pr diff to html
+
+    Args:
+        ctx: the pr diff context
+        theme: the theme of the html
+    """
     template = env.get_template("views/diff.html.jinja")
-    return await template.render_async(issue=issue, theme=theme)
+    return await template.render_async(ctx=ctx, theme=theme)
+
+
+async def issue_opened_to_html(
+    ctx: IssueOpenedContext, theme: Literal["light", "dark"] = "light"
+) -> str:
+    """Render issue or pr opened webhook event to html
+
+    Args:
+        ctx: the issue opened context
+        theme: the theme of the html
+    """
+    template = env.get_template("views/issue-opened.html.jinja")
+    return await template.render_async(ctx=ctx, theme=theme)
+
+
+async def issue_commented_to_html(
+    ctx: IssueCommentedContext, theme: Literal["light", "dark"] = "light"
+) -> str:
+    """Render issue commented webhook event to html
+
+    Args:
+        ctx: the issue commented context
+        theme: the theme of the html
+    """
+    template = env.get_template("views/issue-commented.html.jinja")
+    return await template.render_async(ctx=ctx, theme=theme)
+
+
+async def issue_closed_to_html(
+    ctx: IssueClosedContext, theme: Literal["light", "dark"] = "light"
+) -> str:
+    """Render issue or pr closed webhook event to html
+
+    Args:
+        ctx: the issue closed context
+        theme: the theme of the html
+    """
+    template = env.get_template("views/issue-closed.html.jinja")
+    return await template.render_async(ctx=ctx, theme=theme)
