@@ -1,76 +1,78 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """
 @Author         : yanyongyu
 @Date           : 2022-09-06 09:02:27
 @LastEditors    : yanyongyu
-@LastEditTime   : 2022-09-30 09:12:27
+@LastEditTime   : 2024-03-05 14:38:05
 @Description    : None
 @GitHub         : https://github.com/yanyongyu
 """
+
 __author__ = "yanyongyu"
 
-from nonebot import on_command
-from nonebot.log import logger
-from nonebot.params import Depends
-from githubkit.rest import Installation
+from nonebot import logger, on_command
 from nonebot.plugin import PluginMetadata
 from nonebot.adapters.github import ActionTimeout
-from nonebot.adapters.onebot.v11 import GroupMessageEvent, PrivateMessageEvent
+from githubkit.versions.latest.models import SimpleUser
 
 from src.plugins.github import config
-from src.plugins.github.utils import get_bot
+from src.providers.platform import USER_INFO
+from src.plugins.github.utils import get_github_bot
+from src.plugins.github.helpers import NO_GITHUB_EVENT
 from src.plugins.github.libs.install import create_install_link
-from src.plugins.github.helpers import get_user_info, get_current_user
-
-from .dependencies import get_user_installation
+from src.plugins.github.dependencies import (
+    RUN_WHEN_GROUP,
+    RUN_WHEN_PRIVATE,
+    GITHUB_USER_INSTALLATION,
+)
 
 __plugin_meta__ = PluginMetadata(
     "GitHub APP 集成",
     "集成 GitHub APP 以进行 Issue、PR 相关事件提醒",
-    (
-        "/install: 安装或管理 GitHub APP 集成\n"
-        "/install check: 查看 GitHub APP 集成状态\n"
-        "/install revoke: 撤销 GitHub APP 集成授权"
-    ),
+    "/install: 安装或管理 GitHub APP 集成\n"
+    "/install check: 查看 GitHub APP 集成状态\n"
+    "/install revoke: 撤销 GitHub APP 集成授权",
 )
 
 
-install = on_command("install", priority=config.github_command_priority)
+install = on_command(
+    "install",
+    aliases={"安装"},
+    rule=NO_GITHUB_EVENT,
+    priority=config.github_command_priority,
+    block=True,
+)
 
 
-@install.handle()
-async def handle_group(event: GroupMessageEvent):
+@install.handle(parameterless=(RUN_WHEN_GROUP,))
+async def handle_group():
     await install.finish("请私聊我并使用 /install 命令进行安装或管理")
 
 
-@install.handle()
-async def handle_private(event: PrivateMessageEvent):
-    if info := get_user_info(event):
-        await install.finish("请前往以下链接进行安装或管理：\n" + await create_install_link(info))
-    else:
-        logger.error(f"Unprocessed event type: {type(event)}")
-        await install.finish("内部错误，请尝试私聊我并使用 /install 命令进行安装或管理")
+@install.handle(parameterless=(RUN_WHEN_PRIVATE,))
+async def handle_private(user_info: USER_INFO):
+    await install.finish(
+        "请前往以下链接进行安装或管理：\n" + await create_install_link(user_info)
+    )
 
 
 install_check = on_command(
-    ("install", "check"), priority=config.github_command_priority
+    ("install", "check"),
+    aliases={("安装", "检查")},
+    rule=NO_GITHUB_EVENT,
+    priority=config.github_command_priority,
+    block=True,
 )
 
 
 @install_check.handle()
-async def handle_check(user: None = Depends(get_current_user)):
-    await install_check.finish("你还没有绑定 GitHub 帐号，请使用 /install 进行安装")
-
-
-@install_check.handle()
-async def check_user_installation(
-    installation: Installation = Depends(get_user_installation),
-):
+async def check_user_installation(installation: GITHUB_USER_INSTALLATION):
     # sourcery skip: merge-else-if-into-elif
     repo_selection = installation.repository_selection
     if account := installation.account:
-        gh_user = account.name
+        if isinstance(account, SimpleUser):
+            gh_user = account.name or account.login
+        else:
+            gh_user = account.name or account.slug
         if repo_selection == "selected":
             await install_check.finish(f"{gh_user} 已安装 GitHub APP 并授权了部分仓库")
         else:
@@ -83,21 +85,20 @@ async def check_user_installation(
 
 
 install_revoke = on_command(
-    ("install", "revoke"), priority=config.github_command_priority
+    ("install", "revoke"),
+    aliases={("安装", "撤销")},
+    rule=NO_GITHUB_EVENT,
+    priority=config.github_command_priority,
+    block=True,
 )
 
 
 @install_revoke.handle()
-async def handle_revoke(user: None = Depends(get_current_user)):
-    await install_check.finish("你还没有绑定 GitHub 帐号，请使用 /install 进行安装")
-
-
-@install_revoke.handle()
-async def revoke_user(installation: Installation = Depends(get_user_installation)):
-    bot = get_bot()
+async def revoke_user(installation: GITHUB_USER_INSTALLATION):
+    bot = get_github_bot()
 
     try:
-        await bot.rest.apps.async_delete_installation(installation.id)
+        await bot.rest.apps.async_delete_installation(installation_id=installation.id)
     except ActionTimeout:
         await install_revoke.finish("GitHub API 超时，请稍后再试")
     except Exception as e:
@@ -105,3 +106,5 @@ async def revoke_user(installation: Installation = Depends(get_user_installation
             f"Failed while deleting installation in installation revoke: {e}"
         )
         await install_revoke.finish("未知错误发生，请尝试重试或联系管理员")
+
+    await install_revoke.finish("已移除 GitHub APP 集成")
